Moved the remaining bits from v8wrap to v8context that were context specific.
This moves the parse v8 exception, and execution of v8 scripts.  These now are done within the V8Context class. The idea this change is apart of is to make v8wrap.cc just a export interface for the gocode to use when calling into the v8 library.  A lot of details still need to be added, expecially being able to create instances of registered native objects from javascript. A refactor of how native functions are added is next.